move main files to top level
