+++ package.json
@@ -2,8 +2,8 @@
   "name": "datastore-zipcar",
   "version": "1.1.1",
   "description": "An implementation of a Datastore (https://github.com/ipfs/interface-datastore) for IPLD blocks that operates on ZIP files",
-  "main": "lib/zipcar.js",
-  "browser": "lib/zipcar-browser.js",
+  "main": "zipcar.js",
+  "browser": "zipcar-browser.js",
   "scripts": {
     "lint": "standard",
     "test:node": "hundreds mocha test/test-*.js",
