+++ zipcar.js
@@ -2,10 +2,10 @@ const fs = require('fs')
 const { promisify } = require('util')
 fs.open = promisify(fs.open)
 fs.access = promisify(fs.access)
-const { Reader, EmptyReader, NoWriter, CachingReader, CachingDeferredWriter } = require('./reader-writer-iface')
-const createFileReader = require('./reader-file')
-const createStreamWriter = require('./writer-stream')
-const ZipDatastore = require('./zipdatastore')
+const { Reader, EmptyReader, NoWriter, CachingReader, CachingDeferredWriter } = require('./lib/reader-writer-iface')
+const createFileReader = require('./lib/reader-file')
+const createStreamWriter = require('./lib/writer-stream')
+const ZipDatastore = require('./lib/zipdatastore')
 const { fromBuffer } = require('./zipcar-browser')
 
 async function fromFile (file) {
